requirements.txt
click==7.1.2
cmake==3.17.3
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.18.5
opencv-python-headless==4.2.0.34
Pillow==7.1.2
PyQt5==5.14.2
PyQt5-sip==12.8.0
PyQt5-stubs==5.14.2.2
